feat(SNC-01): create header of application
Create initial configuration to show brand in header of site. In this
commit I installed style-components to help when I use css in the
application to hold it organized and easy to maintain.

Create address area inside of header and badge component because
probaly we will need to use it in other places.

Create cart icon inside of header application, in the future can be
access all items selected to buy in the shop.