feat(SNC-03): create home page of application
Create mockup of home screen to use banner image, title and descrition
to build screen for the home page.

Create context of home page to show title, describe and image of banner.
And I was created a file with all formats to know when the banner is in
mobile, tablet or desktop.

Not just banner, it can be used in other components.

Create badge to show description outside of badge icon as used in Cart
icon, I was created a new component to show a icon with description
beside of it and added in the home screen.