feat(SNC-04): create list of itens in catalog for home page
Create item of catalog and add mock to show how can I show itens for the
user in the home page.

In this commit, we need to create ButtonIcon, BadgeTopic and InputRange
to create all structure necessary for the feature.

Change color of focus and hover components in footer catalog because the
color used was not in the color palette.

Organize styles and components inside of Home feature because previously
organization caused a lot of files with 5 lines of code. Now, the
organization is more clear and easy to understand because all components
is located in same file with "Container" in the final.