feat(SNC-05): create cart page to show products inside of cart
Create the page to show all products inside of cart and it possible to
add product already inside of cart or delete it.

Create section can be show a unique product when user click on cart of
products wants to be buy in the cart page.

I added Roboto font because it not used in the project.

Added colors inside of global theme because I have components need to be
used spefic colors defaulted.

Created "Divider" can be used in the entire project

Updated "ButtonIcon" can be used with and without text.

Added icon "IconThrashPurple" to use in delete item on Cart

Updated "InputRange" to use with max height, it means this component can
be used in small sizes, not specific fixed size.

Added logic can use colors in specific themes used on the
styled-components inside file "hasNeedBlackTextColor".

Resolves: SNC-05-01, SNC-05-02, SNC-05-03.