feat(SNC-05): create section to show product details
Create section can be show a unique product when user click on cart of
products wants to be buy in the cart page.

In this commit I added Roboto font because it not used in the project.

Created "Divider" can be used in the entire project

Updated "ButtonIcon" can be used with and without text.

Added icon "IconThrashPurple" to use in delete item on Cart

Updated "InputRange" to use with max height, it means this component can
be used in small sizes, not specific fixed size.

Added colors inside of global theme because I have components need to be
used spefic colors defaulted.

Added logic can use colors in specific themes used on the
styled-components inside file "hasNeedBlackTextColor".

Resolves: SNC-05-01