feat(SNC-06): create page to register address information
Create form to safe address of user because the user needs to inform the
address to receive the product. The form has the fields: street, number,
complement, neighborhood, city, state and zip code.

We need create components like InputLabelOpcional and
InputTextError to show information about status like error and optional.

We correct a bug inside of InputRange, because this component was not
followed with bottom type "button", for this reason, we added this type
because it submit a form every time I change the quantity in cart.

We added React Hook Forms and yup to validate the form.

Print: https://imgur.com/a/He8dEvL

Resolves: SNC-06, SNC-06-01, SNC-06-02