feat(SNC-07): create header of payment section
Create header of payment section in cart page and add payment icon
with all styles can be used in more than one place, previously used
to in address section, for this reason I created a new file for
these styles global in this feature.

Create a section to select the method of payment in the cart page
between credit, debit or cash.

Add already created icons for each payment method and respective
validation for this component named of CartPaymentOption.

The current schema needs to be updated to include the payment
method because this not have only information about address for
new name: cartSchema.

And new color is added inside of the theme to use when have
a payment method selected.

Resolves: SNC-07-01, SNC-07-02, SNC-07-03