feat(SNC-08): create review page after finish order
Create all configuration about routes and default layout to header page
and all things I need to continue this part of application.

Create two sections, basically we have informations about the order
succeed and image to help a simbolize of delivery is actually comming
to you.

Fill columns of component to show all informations about order, deliver
and payment. Beside we have a image to show the deliver method.

Create border with single colors not specific like the design, because
its not possible to create a border with more than one color in css
without using a image.

And I organized the header of the order informations, putting in the
component was created previously.

I add a prop inside of BadgeDescription to show icon, title and
description text. We also add a component to show custom text like If
you want a text with specific color, size and font.

Resolves: SNC-08-01, SNC-08-02, SNC-08-03, SNC-08-04