feat(SNC-08): fill both columns following the design
Fill columns of component to show all informations about order, deliver
and payment. Beside we have a image to show the deliver method.

In this commit we add a prop inside of BadgeDescription to show icon,
title and description text. We also add a component to show custom text
like If you want a text with specific color, size and font.

Resolves: SNC-08-03