fix: organize code and adjust visual components
Organize code with prettier command and add this in a package.json, it
can possible to run the command `npm run format` or `npm run formatAll`
for all files in the project.

Fixed a height inside of Badge, added fixed height in 38 and correct
formatting of text in header section of the address.

And finally correct the name of CarReviewItem to CartReviewItem.