feat(SNC-09): add products in a local storage to simbolize a cart
Add product in a local storage to simbolize a cart, because we have a
lot of features needs to put a information inside of array to show any
information about the current cart created.

Modified files to add this feature is CartItem, because it file can be
used in more than one feature.

And created a context file to manage all method used to insert, update
and delete information in a specific local storage.