fix: execute "formatAll" and review all these files
Execute "formatAll" to fix all problems in the organization of the code
and review all these files to check if I have any problem or comment
dropped here.

Addionally, I removed the CartItensMocked.ts file, which was not being
used and added a "deletion" of cart when the order is finished.