feat(SNC-75): update responsive of application
Update the responsive design of the application to improve user
experience on various devices. This includes adjustments to layout,
font sizes, and element spacing to ensure a consistent and visually
appealing interface across different screen sizes.

Add a size inside of Devices to represents the new size forgetten
to add in the last commit.